KEYRING: Prepare for the expiration of the old gpg key
The old sign@odex.be key will expire in about a month.
This is a short time for people to update their signature key's before
the new keyring will be put into place.

A new key has been generated and will be used to sign keys starting 3
days before the expiration of the previous one